rename a few more things, clean up a few more apis
