+++ rats-apps/src/python/rats/cli/_plugins.py
@@ -11,10 +11,9 @@ class ClickGroupPlugin(Protocol):
         pass
 
 
-class AttachGroupCommands(ClickGroupPlugin):
-    """
-    When a group is opened, attach a set of commands to it.
-    """
+class AttachClickCommands(ClickGroupPlugin):
+    """When a group is opened, attach a set of commands to it."""
+
     _commands: Iterator[click.Command]
 
     def __init__(self, commands: Iterator[click.Command]) -> None:
