+++ rats-devtools/test/python/rats_test/devtools/test_groups.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock, create_autospec
 import click
 
 from rats.apps import PluginRunner
-from rats.cli import AttachGroupCommands, ClickExecutable
+from rats.cli import AttachClickCommands, ClickExecutable
 
 
 @mock.patch("click.Group")
@@ -34,7 +34,7 @@ def test_command_group_execute_registers_commands(m_click_group: MagicMock):
 
 def test_group_commands_on_group_registration():
     mock_command = create_autospec(spec=click.Command, instance=True)
-    group_commands = AttachGroupCommands(commands=iter([mock_command]))
+    group_commands = AttachClickCommands(commands=iter([mock_command]))
     mock_group = create_autospec(spec=click.Group, instance=True)
     group_commands.on_group_open(mock_group)
     mock_group.add_command.assert_called_once_with(mock_command)
